モデル層の実装
Workspace にドキュメントを追加
ViewModel に処理を移動
