+++ dearx_editor/src/main_window_view_model.rs
@@ -1,7 +1,7 @@
 use std::sync::{Arc, Mutex};
 
 use dearx_edit_model::DearxProject;
-use dearx_workspace::Workspace;
+use dearx_workspace::{DocumentId, Workspace};
 use tauri::Manager;
 
 use crate::PropertyWindowViewModel;
@@ -12,11 +12,8 @@ pub struct MainWindowViewModel {
 }
 
 impl MainWindowViewModel {
-    pub fn new(
-        workspace: Arc<Mutex<Workspace<DearxProject>>>,
-    ) -> Self {
-        let property_window_view_model =
-            PropertyWindowViewModel::new(workspace.clone());
+    pub fn new(id: &DocumentId, workspace: Arc<Mutex<Workspace<DearxProject>>>) -> Self {
+        let property_window_view_model = PropertyWindowViewModel::new(id, workspace.clone());
 
         Self {
             property_window_view_model,
@@ -26,6 +23,14 @@ impl MainWindowViewModel {
     pub fn listen(&self, app: &tauri::App) {
         app.listen_global("front-to-back", |event| {
             println!("Message from frontend: {:?}", event.payload());
+        });
+
+        let app_handle = app.app_handle();
+        std::thread::spawn(move || loop {
+            app_handle
+                .emit_all("back-to-front", "ping frontend".to_string())
+                .unwrap();
+            std::thread::sleep(std::time::Duration::from_secs(1))
         });
     }
 }
